Refactor taxpayer into multiple pages
Add back proper field validation
