+++ src/components/Main.tsx
@@ -2,7 +2,8 @@ import React, { ReactElement } from 'react'
 import { unstable_createMuiStrictModeTheme as createMuiTheme, ThemeProvider } from '@material-ui/core'
 import {
   Switch,
-  Route
+  Route,
+  Redirect
 } from 'react-router-dom'
 import W2JobInfo from './W2JobInfo'
 import CreatePDF from './createPDF'
@@ -10,6 +11,9 @@ import ResponsiveDrawer, { Section } from './ResponsiveDrawer'
 import { PagerButtons, usePager } from './pager'
 import TaxPayerInfo from './TaxPayer'
 import RefundBankAccount from './RefundBankAccount'
+import SpouseAndDependent from './TaxPayer/SpouseAndDependent'
+import ContactInfo from './TaxPayer/ContactInfo'
+import FilingStatusSelect from './TaxPayer/FilingStatus'
 
 const theme = createMuiTheme({
   palette: {
@@ -29,10 +33,14 @@ const theme = createMuiTheme({
 })
 
 const Urls = {
-  taxpayer: '/taxpayerinfo',
+  taxPayer: {
+    root: '/taxpayer',
+    info: '/info',
+    spouseAndDependent: '/spouseAndDependent',
+    filingStatus: '/filingStatus',
+    contactInfo: '/contact'
+  },
   refund: '/refundinfo',
-  employer: '/w2employerinfo',
-  employee: '/w2employeeinfo',
   job: '/w2jobinfo',
   createPdf: '/createpdf'
 }
@@ -41,8 +49,10 @@ const drawerSections: Section[] = [
   {
     title: 'Personal',
     items: [
-      ['Taxpayer Information', Urls.taxpayer],
-      ['Refund Information', Urls.refund]
+      ['Taxpayer Information', Urls.taxPayer.info],
+      ['Spouse and Dependents', Urls.taxPayer.spouseAndDependent],
+      ['Filing Status', Urls.taxPayer.filingStatus],
+      ['Contact Information', Urls.taxPayer.contactInfo]
     ]
   },
   {
@@ -54,6 +64,7 @@ const drawerSections: Section[] = [
   {
     title: 'Results',
     items: [
+      ['Refund Information', Urls.refund],
       ['Review and Print', Urls.createPdf]
     ]
   }
@@ -76,8 +87,18 @@ export default function Main (): ReactElement {
     <ThemeProvider theme={theme}>
       <ResponsiveDrawer sections={drawerSections} />
       <Switch>
-        <Route path={Urls.taxpayer} exact>
+        <Redirect path="/" to={Urls.taxPayer.info} exact />
+        <Route path={Urls.taxPayer.info}>
           <TaxPayerInfo onAdvance={forward} navButtons={firstStepButtons} />
+        </Route>
+        <Route path={Urls.taxPayer.spouseAndDependent}>
+          <SpouseAndDependent onAdvance={forward} navButtons={firstStepButtons} />
+        </Route>
+        <Route path={Urls.taxPayer.contactInfo}>
+          <ContactInfo onAdvance={forward} navButtons={firstStepButtons}/>
+        </Route>
+        <Route path={Urls.taxPayer.filingStatus}>
+          <FilingStatusSelect onAdvance={forward} navButtons={firstStepButtons}/>
         </Route>
         <Route path={Urls.refund} exact>
           <RefundBankAccount onAdvance={forward} navButtons={stepDoneButtons} />
