+++ src/components/TaxPayer/index.tsx
@@ -1,0 +1,2 @@
+import TaxPayer from './TaxPayer'
+export default TaxPayer
